split bgum test aux logic into modules
This commit splits the auxiliary logic for the bubblegum tests into
multiple modules as the amount of code will grow too large for
everything to be in a single place.

It also adds a generic `inner` field to `struct TxBuilder` where tx
specific information can be recorded, and introduces the
`OnSuccessfulTxExec` trait which will be used to implement a mechanism
similar to a callback for tx builder objects such that they can take
actions in response to the tx being successfully executed (this is
primarily so that we can update relevant fields that change - such as
owner and delegate for a leaf post transfer - in an automatic manner,
so manual updates in the test logic are no longer required).
