+++ bubblegum/program/Cargo.toml
@@ -31,6 +31,7 @@ spl-token = { version = "3.5.0", features = ["no-entrypoint"] }
 solana-program-test = "1.10.38"
 solana-sdk = "1.10.38"
 spl-concurrent-merkle-tree = "0.1.0"
+spl-merkle-tree-reference = "0.1.0"
 spl-noop = { version = "0.1.0", features = ["no-entrypoint"] }
 
 [profile.release]
